Remove support for magic link and migrate next auth to app router
