+++ .env.example
@@ -7,20 +7,12 @@
 # When adding additional env variables, the schema in /env/schema.mjs should be updated accordingly
 # Prisma
 DATABASE_URL=your_db_url
-SHADOW_DATABASE_URL=your_db_url
 
 # Next Auth
 # You can generate the secret via 'openssl rand -base64 32' on Linux
 # More info: https://next-auth.js.org/configuration/options#secret
 NEXTAUTH_SECRET=
 NEXTAUTH_URL=http://localhost:3000
-
-# Magic Link
-EMAIL_SERVER_HOST=smtp.gmail.com
-EMAIL_SERVER_PORT=465
-EMAIL_SERVER_USER=example@gmail.com
-EMAIL_SERVER_PASSWORD=app_password_or_acc_password
-EMAIL_FROM=example@gmail.com
 
 # Google Cloud OAuth Credentials
 GOOGLE_ID=
