+++ package.json
@@ -23,7 +23,6 @@
     "clsx": "^2.0.0",
     "copy-to-clipboard": "^3.3.3",
     "csstype": "^3.1.2",
-    "handlebars": "^4.7.8",
     "jotai": "^2.4.0",
     "lodash.debounce": "^4.0.8",
     "nanoid": "^4.0.2",
@@ -33,7 +32,6 @@
     "next-qrcode": "^2.5.1",
     "next-themes": "^0.2.1",
     "nextjs-cors": "^2.1.2",
-    "nodemailer": "^6.9.4",
     "react": "18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "18.2.0",
@@ -50,7 +48,6 @@
     "@types/lodash.debounce": "^4.0.7",
     "@types/next-pwa": "^5.6.4",
     "@types/node": "^20.5.6",
-    "@types/nodemailer": "^6.4.9",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
     "@types/swagger-ui-react": "^4.18.0",
