+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,7 @@
+import { authOptions } from "$lib/auth";
+import NextAuth from "next-auth/next";
+
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
