+++ src/components/AuthModal.tsx
@@ -1,94 +1,15 @@
 import { showAuthModalAtom } from "$store";
-import { isValidEmail } from "$utils/functions";
-import { Dialog, Transition } from "@headlessui/react";
+import { Dialog } from "@headlessui/react";
 import { useAtom } from "jotai";
 import { signIn } from "next-auth/react";
-import { useTheme } from "next-themes";
-import { Fragment, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { toast } from "react-hot-toast";
 import { FcGoogle } from "react-icons/fc";
-import { HiMailOpen } from "react-icons/hi";
-import Button from "./Button";
-import Input from "./Input";
 import Modal from "./Modal";
-
-const Confirm = ({ show = false, email = "" }) => (
-  <Transition appear show={show} as={Fragment}>
-    <div className="fixed inset-0 z-50">
-      <Transition.Child
-        as={Fragment}
-        enter="ease-out duration-300"
-        enterFrom="opacity-0"
-        enterTo="opacity-100"
-        leave="ease-in duration-200"
-        leaveFrom="opacity-100"
-        leaveTo="opacity-0"
-      >
-        <div className="fixed inset-0 bg-white dark:bg-gray-700" />
-      </Transition.Child>
-
-      <Transition.Child
-        as={Fragment}
-        enter="ease-out duration-300"
-        enterFrom="opacity-0 scale-95"
-        enterTo="opacity-100 scale-100"
-        leave="ease-in duration-200"
-        leaveFrom="opacity-100 scale-100"
-        leaveTo="opacity-0 scale-95"
-      >
-        <div className="flex h-full items-center justify-center p-8">
-          <div className="transform overflow-hidden transition-all">
-            <h3 className="text-center text-lg font-medium leading-6">
-              <div className="flex flex-col items-center justify-center space-y-4">
-                <HiMailOpen className="h-12 w-12 shrink-0 text-purple-600 dark:text-lemon-400" />
-              </div>
-              <p className="mt-2 text-2xl font-semibold dark:text-white">
-                Confirm your email
-              </p>
-            </h3>
-
-            <p className="mt-4 text-center text-lg">
-              We emailed a magic link to <strong>{email ?? ""}</strong>.
-              <br />
-              Check your inbox and click the link in the email to login or sign
-              up.
-            </p>
-          </div>
-        </div>
-      </Transition.Child>
-    </div>
-  </Transition>
-);
 
 const AuthModal = () => {
   const [disabled, setDisabled] = useState(false);
-  const [showConfirm, setConfirm] = useState(false);
-  const [email, setEmail] = useState("");
   const [showAuthModal, setShowAuthModal] = useAtom(showAuthModalAtom);
-  const { theme } = useTheme();
-
-  const signInWithEmail = async () => {
-    const toastId = toast.loading("Loading...");
-    try {
-      setDisabled(true);
-      // Perform sign in
-      const res = await signIn("email", {
-        redirect: false,
-        callbackUrl: window.location.href,
-        email,
-      });
-      // Something went wrong
-      if (res?.status !== 200) {
-        throw new Error(res?.error ?? "Unable to sign in");
-      }
-      setConfirm(true);
-      toast.dismiss(toastId);
-    } catch (err) {
-      toast.error("Unable to sign in", { id: toastId });
-    } finally {
-      setDisabled(false);
-    }
-  };
 
   const signInWithGoogle = () => {
     toast.loading("Redirecting...");
@@ -109,8 +30,6 @@ const AuthModal = () => {
       // Wait for 200ms for aniamtion to finish
       setTimeout(() => {
         setDisabled(false);
-        setConfirm(false);
-        setEmail("");
       }, 200);
     }
   }, [showAuthModal]);
@@ -136,37 +55,8 @@ const AuthModal = () => {
             later.
           </Dialog.Description>
 
+          {/* Sign with Google */}
           <div className="mt-10">
-            {/* Sign with email */}
-            <form
-              onSubmit={(e) => {
-                e.preventDefault();
-                void signInWithEmail();
-              }}
-            >
-              <Input
-                name="email"
-                type="email"
-                placeholder="rick@roll.com"
-                disabled={disabled}
-                required
-                onChange={(e) => setEmail(e.target.value)}
-                className="mb-3"
-                variant="modal"
-              />
-              <Button
-                title="Login with Email"
-                variant={theme === "dark" ? "primary" : "tertiary"}
-                className="!m-0 w-full text-base font-normal dark:font-semibold"
-                disabled={disabled || !isValidEmail(email)}
-                type="submit"
-              />
-            </form>
-            <Confirm show={showConfirm} email={email ?? ""} />
-            <p className="my-2 text-center text-sm text-gray-400 dark:text-gray-300">
-              or
-            </p>
-            {/* Sign with Google */}
             <button
               disabled={disabled}
               onClick={signInWithGoogle}
