+++ src/types/next-auth.d.ts
@@ -1,12 +1,21 @@
-import { type DefaultSession } from "next-auth";
+import "next-auth";
+import type { User } from "next-auth";
+import "next-auth/jwt";
+
+type UserId = string;
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    id: UserId;
+    username?: string | null;
+  }
+}
 
 declare module "next-auth" {
-  /**
-   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
-   */
   interface Session {
-    user?: {
-      id: string;
-    } & DefaultSession["user"];
+    user: User & {
+      id: UserId;
+      username?: string | null;
+    };
   }
 }
